Implement Dagwood with Lcl (#3)
* wip

* WIP

* Whitespace cleanup

* Refactoring

* WIP

* Cleanup

* Proper Project-level attribute handling

* Initial orchestration with spawn.h

* WIP

* Cleanup and posix_spawn working

* Cleanup

* WIP

* Cleanup DSL

* More consistent logging

* Initial CLI

* Cross-builds working

* Checkpoint

Need to start testing now. Core functions work but need to setup
Dagwood files in a test directory and exercise all the weird edge
cases, especially with cross-builds.

* Individual task running appears to work

* List tasks

Next piece of work:

Right now, we're having C call some Tcl. It should be the reverse. We
should, in principle, be able to de-sugar an entire Dagwood operation
into a series of Tcl commands. This will eventually make REPL-driven
development easier.

* Basic REPL

* Fixed a bunch of memory leaks

* Whitespace cleanup

* Cleanup

* More memory leaks

* Simple Project e2e tests

* More cleanup

* Awesome test

* wip

* Configurable shell

* Cleanup

* Add better compile opts, fix many warnings and cleanup

* Apply default shell everywhere

* Add better documentation

* More cleanup

* Move Tcl file to have consistent naming with C header

* Cleanup some unused functions

* Formatting

* Checkpoint before formatting

* Format

* Cleanup

* Format

* Combine format and build

* Better documentation

* Start to clean up interpreter API

* Fix the rest of the memory leaks

* Add finalize for good measure

* More consistent ownership semantics

klib data structures now shouldn't explicitly own their elements. The
containing structs should take ownership. This is slightly more
messy (arguably), however, this project is predicated on lots of
arrays (or klib `vec`s) of pointers in order for the graph algorithms
like Kahn's to work efficiently.

This is maybe not ideal, but graph theoretic problems are amenable to
passing pointers around.

tl;dr the stance going forward: klib data structures won't "own" or
otherwise decide the lifetimes of their elements. The "owning"
structs (SoA) can clear the elements manually, and deliberately.

* Fixed stack upleveling in DSL.

Now things like `task` and `task-in` are equivalent. There isn't a
need to remember to do `task-in` within a proc.

* Removed some unnecessary attributes, minor cleanup

* Remove timeout, remove unused tests

This is a design decision, but (arguably) timeout is not necessary at
the task-level. A caller could run the standard coreutil command

```
timeout 5 dagwood
```

Which would semantically be the same.

This avoids us having to implement timeouts ontop of individual
`posix_spawn`s -- which is not exactly trivial -- and respects the
Unix philosophy.

* At spawngroup and bind to Dagwood pid

* Forward signals to child processes

* Better documentation, improve logging

* More consistent naming

* Initial Graphviz

* Better Graphviz DOT ordering

* Fix test

* Initial README

* Add gitattributes

* Highlight any Dagwood* file

* More README

* Minor addition to FAQ

* WIP

* TODO list and moved tests directory

* Minor doc/todo update

* Complete rewrite of import/substitution

* Checkpoint commit

I'm going to do some weird things after this.

Checkpointing here to save work

* Big redesign part 1

* Fix memory leak, reorganize under dagwood.h

* Update Dagwood project file to newer variant

* Updated the documentated Dagwood file.

* Update README

* Move interpreter to core

* Fix list_all command

* Add run task and force cli option

* Fix DOT command

* Add dry-run option

* Initial command

* Add file-normalizing Tcl proc

* Stop tracking the generated dagwood_stdlib.h file

* Fixes to working directory handling in imported projects

* Fix whitespace inconsistencies

* Simplified folder structure

* Initial Jim Tcl refactor -- very WIP

* Jim Tcl refactor -- still WIP

* Checkpoint (massive refactor incoming)

* Fixed majority of the major issues, no memory leaks

* Various fixes, propagate task ID now for stable identifier

* Checkpoint

* Checkpoint before converting to Lcl

* Initial conversion to Lcl

* Remove klib -- pure C89 in-house data structures

* Fix various test issues and DAG resolution

* Add build args

* Update Lcl and pull in lcl-io module by default